+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [0.4.1] - 2024-05-17
+
+- fix timestamps. Thanks @mw10013
+
+## [0.4.0] - 2024-05-14
+
+- monotonic timers. Thanks @mw10013
+
 ## [0.3.2] - 2024-04-18
 
 - fix: add a fallback dataset